Revert a couple of renamings
