+++ specs/test_formats/operations/README.md
@@ -14,7 +14,7 @@ post: BeaconState                      -- state after applying the operation. No
 
 ## Condition
 
-A handler of the `operations` test-runner should process these cases, 
+A handler of the `operations` test-runner should process these cases,
  calling the corresponding processing implementation.
 
 Operations:
@@ -23,13 +23,13 @@ Operations:
 |-------------------------|----------------------|----------------------|--------------------------------------------------------|
 | `attestation`           | `Attestation`        | `attestation`        | `process_attestation(state, attestation)`                  |
 | `attester_slashing`     | `AttesterSlashing`   | `attester_slashing`  | `process_attester_slashing(state, attester_slashing)`            |
-| `block_header`          | `Block`              | `block`              | `process_block_header(state, block)`                   |
+| `parent_block_header`          | `Block`              | `block`              | `process_block_header(state, block)`                   |
 | `deposit`               | `Deposit`            | `deposit`            | `process_deposit(state, deposit)`                      |
 | `proposer_slashing`     | `ProposerSlashing`   | `proposer_slashing`  | `process_proposer_slashing(state, proposer_slashing)`  |
 | `transfer`              | `Transfer`           | `transfer`           | `process_transfer(state, transfer)`                    |
 | `voluntary_exit`        | `VoluntaryExit`      | `voluntary_exit`     | `process_voluntary_exit(state, voluntary_exit)`        |
 
-Note that `block_header` is not strictly an operation (and is a full `Block`), but processed in the same manner, and hence included here. 
+Note that `parent_block_header` is not strictly an operation (and is a full `Block`), but processed in the same manner, and hence included here.
 
 The resulting state should match the expected `post` state, or if the `post` state is left blank,
  the handler should reject the input operation as invalid.
