+++ test_generators/operations/main.py
@@ -42,8 +42,8 @@ if __name__ == "__main__":
         create_suite('attestation',       'mainnet', lambda: generate_from_tests(test_process_attestation)),
         create_suite('attester_slashing', 'minimal', lambda: generate_from_tests(test_process_attester_slashing)),
         create_suite('attester_slashing', 'mainnet', lambda: generate_from_tests(test_process_attester_slashing)),
-        create_suite('block_header',      'minimal', lambda: generate_from_tests(test_process_block_header)),
-        create_suite('block_header',      'mainnet', lambda: generate_from_tests(test_process_block_header)),
+        create_suite('parent_block_header',      'minimal', lambda: generate_from_tests(test_process_block_header)),
+        create_suite('parent_block_header',      'mainnet', lambda: generate_from_tests(test_process_block_header)),
         create_suite('deposit',           'minimal', lambda: generate_from_tests(test_process_deposit)),
         create_suite('deposit',           'mainnet', lambda: generate_from_tests(test_process_deposit)),
         create_suite('proposer_slashing', 'minimal', lambda: generate_from_tests(test_process_proposer_slashing)),
